Fix story room exports and restore caretaker reminder dialogue
