+++ lib/story/rooms/chapter1/index.ts
@@ -1,0 +1,9 @@
+export { buildBluffPassageway } from "./bluff_passage";
+export { buildBluffCaves } from "./bluff_caves";
+export { buildBluffSerpentDen } from "./bluff_serpent_den";
+export { buildEntranceHall } from "./entrance_hall";
+export { buildAscentCorridor } from "./ascent_corridor";
+export { buildSanctum } from "./sanctum";
+export { buildOutdoorClearing } from "./outdoor_clearing";
+export { buildOutdoorHouse } from "./outdoor_house";
+export { buildTorchTown } from "./torch_town";
