+++ lib/story/story_mode.ts
@@ -8,364 +8,28 @@ import {
   createCheckpointSnapshot,
   findPlayerPosition,
   isWithinBounds,
+  FLOOR,
+  WALL,
 } from "../map";
 import type { EnvironmentId } from "../environment";
 import { Enemy, EnemyState, rehydrateEnemies, type PlainEnemy } from "../enemy";
 import { NPC, rehydrateNPCs, serializeNPCs } from "../npc";
 import { createInitialStoryFlags } from "./event_registry";
-
-const FLOOR = 0;
-const WALL = 1;
+import {
+  buildAscentCorridor,
+  buildBluffCaves,
+  buildBluffPassageway,
+  buildBluffSerpentDen,
+  buildEntranceHall,
+  buildOutdoorClearing,
+  buildOutdoorHouse,
+  buildSanctum,
+  buildTorchTown,
+} from "./rooms/chapter1";
+import type { StoryRoom } from "./rooms/types";
 
 function cloneMapData(mapData: MapData): MapData {
   return JSON.parse(JSON.stringify(mapData)) as MapData;
-}
-
-// 10x40 bluff passageway room with right-side lower opening
-function buildBluffPassageway(): StoryRoom {
-  const height = 10;
-  const width = 40;
-  const tiles: number[][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => [] as number[])
-  );
-
-  // Carve interior floor
-  for (let y = 1; y < height - 1; y++) {
-    for (let x = 1; x < width - 1; x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  // Create a static interior pattern: a few wall bands and pillars
-  const bands = [3, 6];
-  for (const by of bands) {
-    for (let x = 5; x < width - 5; x++) {
-      if (x % 6 !== 0) continue; // sparse band posts
-      tiles[by][x] = WALL;
-      subtypes[by][x] = [];
-    }
-  }
-  // Pillars near the middle
-  const midY = Math.floor(height / 2);
-  const midX = Math.floor(width / 2);
-  const pillarSpots: Array<[number, number]> = [
-    [midY - 1, midX - 4],
-    [midY + 1, midX - 2],
-    [midY, midX],
-    [midY - 1, midX + 3],
-    [midY + 1, midX + 6],
-  ];
-  for (const [py, px] of pillarSpots) {
-    if (tiles[py]?.[px] !== undefined) {
-      tiles[py][px] = WALL;
-      subtypes[py][px] = [];
-    }
-  }
-
-  // Opening on right wall near bottom
-  const openY = height - 3; // two tiles up from bottom
-  tiles[openY][width - 1] = FLOOR;
-  subtypes[openY][width - 1] = [TileSubtype.ROOM_TRANSITION];
-
-  // Opening at the top-left interior that leads deeper into the bluff
-  const bluffCaveEntry: [number, number] = [1, 3];
-  tiles[bluffCaveEntry[0]][bluffCaveEntry[1]] = FLOOR;
-  subtypes[bluffCaveEntry[0]][bluffCaveEntry[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-
-  const entryPoint: [number, number] = [openY, width - 2]; // step inside from right
-  const entryFromNext: [number, number] = [openY, width - 3];
-  const transitionToPrevious: [number, number] = [openY, width - 1];
-
-  // Enemies: 5 snakes and 1 goblin near center
-  const enemies: Enemy[] = [];
-  const enemySpots: Array<[number, number, 'snake' | 'goblin']> = [
-    [midY, midX - 2, 'snake'],
-    [midY - 1, midX, 'snake'],
-    [midY + 1, midX + 1, 'snake'],
-    [midY, midX + 3, 'snake'],
-    [midY + 2, midX - 1, 'snake'],
-    [midY, midX, 'goblin'],
-  ];
-  for (const [y, x, kind] of enemySpots) {
-    if (tiles[y]?.[x] === FLOOR) {
-      const e = new Enemy({ y, x });
-      e.kind = kind;
-      enemies.push(e);
-    }
-  }
-
-  // NPC boy on the far left, with a goblin near him
-  const boyY = midY;
-  const boyX = 1; // far left inside wall
-  const boy = new NPC({
-    id: "npc-sanctum-boy",
-    name: "Sanctum Acolyte",
-    sprite: "/images/npcs/boy-3.png",
-    y: boyY,
-    x: boyX,
-    facing: Direction.RIGHT,
-    canMove: false,
-    interactionHooks: [
-      {
-        id: "boy-greet",
-        type: "dialogue",
-        description: "Check on the boy",
-        payload: { dialogueId: "librarian-default" },
-      },
-    ],
-    actions: ["talk"],
-  });
-  const guardGoblinPos: [number, number] = [Math.max(1, boyY - 1), Math.min(width - 3, boyX + 3)];
-  if (tiles[guardGoblinPos[0]]?.[guardGoblinPos[1]] === FLOOR) {
-    const guard = new Enemy({ y: guardGoblinPos[0], x: guardGoblinPos[1] });
-    guard.kind = 'goblin';
-    enemies.push(guard);
-  }
-
-  return {
-    id: "story-bluff-passage" as RoomId,
-    mapData: { tiles, subtypes, environment: "outdoor" },
-    entryPoint,
-    entryFromNext,
-    transitionToPrevious,
-    enemies,
-    npcs: [boy],
-  };
-}
-
-function buildBluffCaves(): StoryRoom {
-  const height = 20;
-  const width = 40;
-  const tiles: number[][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => [] as number[])
-  );
-
-  const ensureFloor = (y: number, x: number) => {
-    if (y <= 0 || y >= height - 1 || x <= 0 || x >= width - 1) return;
-    tiles[y][x] = FLOOR;
-    if (!subtypes[y][x]) subtypes[y][x] = [];
-  };
-
-  const carveBlock = (cy: number, cx: number, radius = 0) => {
-    for (let dy = -radius; dy <= radius; dy++) {
-      for (let dx = -radius; dx <= radius; dx++) {
-        ensureFloor(cy + dy, cx + dx);
-      }
-    }
-  };
-
-  const carvePath = (
-    points: Array<[number, number]>,
-    radius = 0
-  ) => {
-    if (points.length === 0) return;
-    let [cy, cx] = points[0];
-    carveBlock(cy, cx, radius);
-    for (let i = 1; i < points.length; i++) {
-      const [ny, nx] = points[i];
-      const stepY = Math.sign(ny - cy);
-      const stepX = Math.sign(nx - cx);
-      while (cy !== ny || cx !== nx) {
-        if (cx !== nx) {
-          cx += stepX;
-        } else if (cy !== ny) {
-          cy += stepY;
-        }
-        carveBlock(cy, cx, radius);
-      }
-    }
-  };
-
-  const mainPath: Array<[number, number]> = [
-    [1, 1],
-    [1, 15],
-    [4, 15],
-    [4, 4],
-    [8, 4],
-    [8, 22],
-    [5, 22],
-    [5, 30],
-    [14, 30],
-    [14, 18],
-    [10, 18],
-    [10, 38],
-  ];
-  carvePath(mainPath, 1);
-
-  const branches: Array<{ path: Array<[number, number]>; radius?: number }> = [
-    { path: [
-        [4, 9],
-        [6, 9],
-        [6, 13],
-      ], radius: 0 },
-    { path: [
-        [8, 12],
-        [12, 12],
-        [12, 8],
-      ], radius: 0 },
-    { path: [
-        [10, 26],
-        [16, 26],
-        [16, 34],
-      ], radius: 1 },
-  ];
-  for (const branch of branches) {
-    carvePath(branch.path, branch.radius ?? 0);
-  }
-
-  // Create a few wider pockets off the main path for a winding feel
-  const pockets: Array<[number, number, number]> = [
-    [5, 7, 1],
-    [9, 20, 1],
-    [13, 24, 1],
-  ];
-  for (const [py, px, radius] of pockets) {
-    carveBlock(py, px, radius);
-  }
-
-  const entryDoorY = 1;
-  tiles[entryDoorY][0] = FLOOR;
-  subtypes[entryDoorY][0] = [TileSubtype.ROOM_TRANSITION];
-  const entryPoint: [number, number] = [entryDoorY, 1];
-  const transitionToPrevious: [number, number] = [entryDoorY, 0];
-
-  const exitDoorY = 10;
-  tiles[exitDoorY][width - 1] = FLOOR;
-  subtypes[exitDoorY][width - 1] = [TileSubtype.ROOM_TRANSITION];
-  const transitionToNext: [number, number] = [exitDoorY, width - 1];
-  const entryFromNext: [number, number] = [exitDoorY, width - 2];
-
-  const torchPositions: Array<[number, number]> = [
-    [exitDoorY - 1, width - 2],
-    [exitDoorY + 1, width - 2],
-  ];
-  for (const [ty, tx] of torchPositions) {
-    if (ty > 0 && ty < height - 1 && tx > 0 && tx < width - 1) {
-      ensureFloor(ty, tx);
-      subtypes[ty][tx] = [TileSubtype.WALL_TORCH];
-    }
-  }
-
-  const enemies: Enemy[] = [];
-  const snakeSpots: Array<[number, number]> = [
-    [4, 12],
-    [8, 21],
-    [13, 23],
-  ];
-  for (const [y, x] of snakeSpots) {
-    if (tiles[y]?.[x] === FLOOR) {
-      const snake = new Enemy({ y, x });
-      snake.kind = "snake";
-      enemies.push(snake);
-    }
-  }
-
-  return {
-    id: "story-bluff-caves" as RoomId,
-    mapData: { tiles, subtypes, environment: "cave" },
-    entryPoint,
-    entryFromNext,
-    transitionToPrevious,
-    transitionToNext,
-    enemies,
-  };
-}
-
-function buildBluffSerpentDen(): StoryRoom {
-  const size = 12;
-  const tiles: number[][] = Array.from({ length: size }, () =>
-    Array.from({ length: size }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: size }, () =>
-    Array.from({ length: size }, () => [] as number[])
-  );
-
-  for (let y = 1; y < size - 1; y++) {
-    for (let x = 1; x < size - 1; x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  const entryY = Math.floor(size / 2);
-  const entryPoint: [number, number] = [entryY, 1];
-  const transitionToPrevious: [number, number] = [entryY, 0];
-  tiles[entryY][0] = FLOOR;
-  subtypes[entryY][0] = [TileSubtype.ROOM_TRANSITION];
-
-  const entryFromNext: [number, number] = [entryY, 2];
-
-  const entranceTorches: Array<[number, number]> = [
-    [entryY - 1, 1],
-    [entryY + 1, 1],
-  ];
-  for (const [ty, tx] of entranceTorches) {
-    if (ty > 0 && ty < size - 1) {
-      subtypes[ty][tx] = [TileSubtype.WALL_TORCH];
-    }
-  }
-
-  const npcY = entryY;
-  const npcX = size - 2;
-  const coiledSnake = new NPC({
-    id: "npc-bluff-coiled-snake",
-    name: "Coiled Snake",
-    sprite: "/images/enemies/snake-coiled-right.png",
-    y: npcY,
-    x: npcX,
-    facing: Direction.LEFT,
-    canMove: false,
-    interactionHooks: [
-      {
-        id: "coiled-snake-greeting",
-        type: "dialogue",
-        description: "Speak with the coiled snake",
-        payload: { dialogueId: "bluff-coiled-snake" },
-      },
-    ],
-    actions: ["talk"],
-  });
-
-  const enemies: Enemy[] = [];
-  const snakePositions: Array<[number, number]> = [];
-  for (let y = 1; y < size - 1 && snakePositions.length < 30; y++) {
-    for (let x = 1; x < size - 1 && snakePositions.length < 30; x++) {
-      if (x <= 2) continue;
-      if (
-        (y === entryPoint[0] && x <= entryPoint[1] + 2) ||
-        (y === npcY && x === npcX)
-      ) {
-        continue;
-      }
-      snakePositions.push([y, x]);
-    }
-  }
-
-  for (let i = 0; i < 30 && i < snakePositions.length; i++) {
-    const [y, x] = snakePositions[i];
-    if (tiles[y]?.[x] === FLOOR) {
-      const snake = new Enemy({ y, x });
-      snake.kind = "snake";
-      enemies.push(snake);
-    }
-  }
-
-  return {
-    id: "story-bluff-serpent-den" as RoomId,
-    mapData: { tiles, subtypes, environment: "cave" },
-    entryPoint,
-    entryFromNext,
-    transitionToPrevious,
-    enemies,
-    npcs: [coiledSnake],
-  };
 }
 
 function withoutPlayer(mapData: MapData): MapData {
@@ -424,819 +88,11 @@ function cloneNPCs(npcs?: NPC[]): NPC[] {
   return rehydrateNPCs(plain);
 }
 
-type StoryRoomLink = {
-  roomId: RoomId;
-  position: [number, number];
-  targetEntryPoint?: [number, number];
-  returnEntryPoint?: [number, number];
-};
-
-type StoryRoom = {
-  id: RoomId;
-  mapData: MapData;
-  entryPoint: [number, number];
-  returnEntryPoint?: [number, number];
-  entryFromNext?: [number, number];
-  transitionToNext?: [number, number];
-  transitionToPrevious?: [number, number];
-  enemies?: Enemy[];
-  npcs?: NPC[];
-  potOverrides?: Record<string, TileSubtype.FOOD | TileSubtype.MED>;
-  metadata?: Record<string, unknown>;
-  otherTransitions?: StoryRoomLink[];
-};
-
-function buildEntranceHall(): StoryRoom {
-  const HALL_LENGTH = 20;
-  const HALL_WIDTH = 3;
-  const height = HALL_WIDTH + 2;
-  const width = HALL_LENGTH + 2;
-
-  const tiles: number[][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => [] as number[])
-  );
-
-  for (let y = 1; y <= HALL_WIDTH; y++) {
-    for (let x = 1; x <= HALL_LENGTH; x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  const midRow = 1 + Math.floor(HALL_WIDTH / 2);
-  const entryPoint: [number, number] = [midRow, 2];
-  const returnEntryPoint: [number, number] = [
-    midRow,
-    Math.max(2, HALL_LENGTH - 1),
-  ];
-  const rightWallX = width - 1;
-  for (let y = 1; y <= HALL_WIDTH; y++) {
-    tiles[y][rightWallX] = FLOOR;
-    subtypes[y][rightWallX] = [];
-  }
-  const transitionToNext: [number, number] = [midRow, rightWallX];
-  subtypes[transitionToNext[0]][transitionToNext[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-
-  const topWall = 0;
-  const bottomWall = height - 1;
-  const torchInterval = 6;
-  for (let offset = 2; offset <= HALL_LENGTH; offset += torchInterval) {
-    const torchX = offset;
-    if (subtypes[topWall][torchX].length === 0) {
-      subtypes[topWall][torchX] = [TileSubtype.WALL_TORCH];
-    }
-    if (subtypes[bottomWall][torchX].length === 0) {
-      subtypes[bottomWall][torchX] = [TileSubtype.WALL_TORCH];
-    }
-  }
-
-  const potOverrides: Record<string, TileSubtype.FOOD | TileSubtype.MED> = {};
-  const potPositions: Array<[number, number]> = [
-    [1, Math.max(3, Math.floor(width / 4))],
-    [1, Math.min(width - 2, Math.floor((width * 3) / 4))],
-  ];
-  for (const [py, px] of potPositions) {
-    if (tiles[py]?.[px] === FLOOR) {
-      subtypes[py][px] = [TileSubtype.POT];
-      potOverrides[`${py},${px}`] = TileSubtype.FOOD;
-    }
-  }
-
-  // No mentor in the entrance hall; elder is placed outside in the clearing
-  const npcs: NPC[] = [];
-
-  return {
-    id: "story-hall-entrance",
-    mapData: { tiles, subtypes, environment: "cave" },
-    entryPoint,
-    returnEntryPoint,
-    transitionToNext,
-    potOverrides,
-    npcs,
-  };
-}
-
-function buildAscentCorridor(): StoryRoom {
-  const VERTICAL_STEPS = 10;
-  const CORRIDOR_WIDTH = 3;
-  const height = VERTICAL_STEPS + 4;
-  const width = CORRIDOR_WIDTH + 4;
-
-  const tiles: number[][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => [] as number[])
-  );
-
-  const bottom = height - 2;
-  const top = bottom - (VERTICAL_STEPS - 1);
-  for (let y = top; y <= bottom; y++) {
-    for (let x = 2; x <= 2 + (CORRIDOR_WIDTH - 1); x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  const entryRow = bottom - 1;
-  for (let y = entryRow - 1; y <= entryRow + 1; y++) {
-    for (let x = 1; x <= 2 + (CORRIDOR_WIDTH - 1); x++) {
-      if (y >= top && y <= bottom) {
-        tiles[y][x] = FLOOR;
-      }
-    }
-  }
-
-  for (let y = entryRow - 1; y <= entryRow + 1; y++) {
-    if (y >= top && y <= bottom) {
-      tiles[y][0] = FLOOR;
-      subtypes[y][0] = [];
-    }
-  }
-
-  const hallwayX = 3;
-  const hallwayLength = 3;
-  const hallwayStartY = Math.max(0, top - (hallwayLength - 1));
-  for (let y = top; y >= hallwayStartY; y--) {
-    for (let x = 2; x <= 4; x++) {
-      if (x === hallwayX) {
-        tiles[y][x] = FLOOR;
-        if (!subtypes[y][x]) subtypes[y][x] = [];
-      } else {
-        tiles[y][x] = WALL;
-        subtypes[y][x] = [];
-      }
-    }
-  }
-
-  const openExitY = hallwayStartY - 1;
-  if (openExitY >= 0 && tiles[openExitY]) {
-    tiles[openExitY][hallwayX] = FLOOR;
-    subtypes[openExitY][hallwayX] = subtypes[openExitY][hallwayX] ?? [];
-  }
-
-  const transitionToPrevious: [number, number] = [entryRow, 0];
-  subtypes[transitionToPrevious[0]][transitionToPrevious[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-
-  const entryPoint: [number, number] = [entryRow, 1];
-  const entryFromNext: [number, number] = [
-    Math.min(bottom, hallwayStartY + 1),
-    hallwayX,
-  ];
-  const transitionToNext: [number, number] = [hallwayStartY, hallwayX];
-  subtypes[transitionToNext[0]][transitionToNext[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-
-  const potOverrides: Record<string, TileSubtype.FOOD | TileSubtype.MED> = {};
-  const midSectionY = Math.max(
-    hallwayStartY + 1,
-    Math.floor((top + bottom) / 2)
-  );
-  const potPositions: Array<[number, number]> = [
-    [midSectionY, 2],
-    [midSectionY, 4],
-  ];
-  for (const [py, px] of potPositions) {
-    if (tiles[py]?.[px] === FLOOR) {
-      subtypes[py][px] = [TileSubtype.POT];
-      potOverrides[`${py},${px}`] = TileSubtype.FOOD;
-    }
-  }
-
-  const torchColumns = [1, width - 2];
-  for (let y = bottom; y >= top; y -= 4) {
-    for (const col of torchColumns) {
-      if (tiles[y][col] === WALL) {
-        subtypes[y][col] = [TileSubtype.WALL_TORCH];
-      }
-    }
-  }
-
-  return {
-    id: "story-ascent",
-    mapData: { tiles, subtypes, environment: "cave" },
-    entryPoint,
-    returnEntryPoint: [entryRow, 4],
-    entryFromNext,
-    transitionToNext,
-    transitionToPrevious,
-    potOverrides,
-  };
-}
-
-function buildSanctum(): StoryRoom {
-  const INNER_SIZE = 10;
-  const height = INNER_SIZE + 2;
-  const width = INNER_SIZE + 2;
-
-  const tiles: number[][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => [] as number[])
-  );
-
-  for (let y = 1; y <= INNER_SIZE; y++) {
-    for (let x = 1; x <= INNER_SIZE; x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  const entryX = Math.floor(width / 2);
-  const transitionY = height - 2;
-  const hallwayLength = 3;
-  const hallwayTopY = Math.max(1, transitionY - (hallwayLength - 1));
-  for (let y = transitionY; y >= hallwayTopY; y--) {
-    for (let x = 1; x <= INNER_SIZE; x++) {
-      if (x === entryX) {
-        tiles[y][x] = FLOOR;
-        if (!subtypes[y][x]) subtypes[y][x] = [];
-      } else {
-        tiles[y][x] = WALL;
-        subtypes[y][x] = [];
-      }
-    }
-  }
-
-  const entryY = Math.max(1, hallwayTopY - 1);
-  const entryPoint: [number, number] = [entryY, entryX];
-  const transitionToPrevious: [number, number] = [transitionY, entryX];
-  subtypes[transitionToPrevious[0]][transitionToPrevious[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-
-  const openBottomY = transitionY + 1;
-  if (openBottomY < tiles.length) {
-    tiles[openBottomY][entryX] = FLOOR;
-    subtypes[openBottomY][entryX] = subtypes[openBottomY][entryX] ?? [];
-  }
-
-  const potOverrides: Record<string, TileSubtype.FOOD | TileSubtype.MED> = {};
-  const potPositions: Array<[number, number]> = [
-    [entryY - 2, entryX - 3],
-    [entryY - 2, entryX + 3],
-  ];
-  for (const [py, px] of potPositions) {
-    if (tiles[py]?.[px] === FLOOR) {
-      subtypes[py][px] = [TileSubtype.POT];
-      potOverrides[`${py},${px}`] = TileSubtype.MED;
-    }
-  }
-
-  const topRowY = 1;
-  for (let x = 1; x <= INNER_SIZE; x++) {
-    if (x === entryX) {
-      continue;
-    }
-    subtypes[topRowY][x] = [];
-    if (potOverrides[`${topRowY},${x}`]) {
-      delete potOverrides[`${topRowY},${x}`];
-    }
-  }
-  tiles[0][entryX] = FLOOR;
-  const transitionToNext: [number, number] = [0, entryX];
-  subtypes[transitionToNext[0]][transitionToNext[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-
-  const entryFromNext: [number, number] = [transitionToNext[0] + 1, entryX];
-
-  const snakes: Enemy[] = [];
-  const snakeA = new Enemy({ y: entryY - 3, x: entryX - 2 });
-  snakeA.kind = "snake";
-  snakes.push(snakeA);
-  const snakeB = new Enemy({ y: entryY - 1, x: entryX + 2 });
-  snakeB.kind = "snake";
-  snakes.push(snakeB);
-
-  const torchRow = 1;
-  const torchCols = [entryX - 1, entryX + 1].filter(
-    (x) => x >= 1 && x <= INNER_SIZE
-  );
-  for (const x of torchCols) {
-    if (tiles[torchRow][x] === WALL) {
-      subtypes[torchRow][x] = [TileSubtype.WALL_TORCH];
-    }
-  }
-
-  return {
-    id: "story-sanctum",
-    mapData: { tiles, subtypes, environment: "cave" },
-    entryPoint,
-    transitionToPrevious,
-    entryFromNext,
-    transitionToNext,
-    enemies: snakes,
-    potOverrides,
-  };
-}
-
-function buildOutdoorWorld(): StoryRoom {
-  const OUTER_WIDTH = 12;
-  const OUTER_HEIGHT = 20;
-  const width = OUTER_WIDTH + 2;
-  const height = OUTER_HEIGHT + 2;
-
-  const tiles: number[][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: height }, () =>
-    Array.from({ length: width }, () => [] as number[])
-  );
-
-  for (let y = 1; y <= OUTER_HEIGHT; y++) {
-    for (let x = 1; x <= OUTER_WIDTH; x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  const entryX = 1 + Math.floor((OUTER_WIDTH - 1) / 2);
-  const entryY = height - 4;
-  const bottomOpeningY = height - 1;
-  tiles[bottomOpeningY][entryX] = FLOOR;
-  subtypes[bottomOpeningY][entryX] = [TileSubtype.ROOM_TRANSITION];
-
-  const entryPoint: [number, number] = [bottomOpeningY - 1, entryX];
-  const transitionToPrevious: [number, number] = [bottomOpeningY, entryX];
-
-  const checkpointY = Math.max(1, entryPoint[0] - 7);
-  const checkpointX = entryPoint[1];
-  if (tiles[checkpointY]?.[checkpointX] === FLOOR) {
-    subtypes[checkpointY][checkpointX] = [TileSubtype.CHECKPOINT];
-  }
-
-  // Add small clusters of rocks to give the outdoor space some structure
-  const featureSpots: Array<[number, number, number[]]> = [
-    [entryY - 4, entryX - 4, [TileSubtype.ROCK]],
-    [entryY - 6, entryX + 3, [TileSubtype.ROCK]],
-    [entryY + 2, entryX - 3, [TileSubtype.ROCK]],
-  ];
-  for (const [fy, fx, values] of featureSpots) {
-    if (tiles[fy]?.[fx] === FLOOR && subtypes[fy][fx].length === 0) {
-      subtypes[fy][fx] = values.slice();
-    }
-  }
-
-  const houseTopY = entryY - 12;
-  const houseHeight = 3;
-  const houseLeftX = OUTER_WIDTH - 4;
-
-  for (let y = 0; y < houseHeight; y++) {
-    const row = houseTopY + y;
-    for (let x = 0; x < 3; x++) {
-      const col = houseLeftX + x;
-      if (tiles[row]?.[col] !== undefined) {
-        tiles[row][col] = WALL;
-        subtypes[row][col] = [];
-      }
-    }
-  }
-
-  const doorY = houseTopY + houseHeight - 1;
-  const doorX = houseLeftX + 1;
-  tiles[doorY][doorX] = WALL;
-  subtypes[doorY][doorX] = [TileSubtype.DOOR, TileSubtype.ROOM_TRANSITION];
-
-  const exteriorDoorY = doorY + 1;
-  if (tiles[exteriorDoorY]?.[doorX] === WALL) {
-    tiles[exteriorDoorY][doorX] = FLOOR;
-    subtypes[exteriorDoorY][doorX] = [];
-  }
-
-  const torchTownTransition: [number, number] = [0, width - 2];
-  tiles[torchTownTransition[0]][torchTownTransition[1]] = FLOOR;
-  subtypes[torchTownTransition[0]][torchTownTransition[1]] = [
-    TileSubtype.ROOM_TRANSITION,
-  ];
-  const torchTownReturn: [number, number] = [
-    Math.min(height - 2, torchTownTransition[0] + 1),
-    torchTownTransition[1],
-  ];
-  if (tiles[torchTownReturn[0]]?.[torchTownReturn[1]] === WALL) {
-    tiles[torchTownReturn[0]][torchTownReturn[1]] = FLOOR;
-    subtypes[torchTownReturn[0]][torchTownReturn[1]] = [];
-  }
-
-  // Place Elder Rowan just outside the cave entrance: 3 tiles up and 3 to the right of the mouth
-  const elderY = Math.max(1, entryPoint[0] - 3);
-  const elderX = Math.min(width - 2, entryPoint[1] + 3);
-  const elder = new NPC({
-    id: "npc-elder-rowan",
-    name: "Elder Rowan",
-    sprite: "/images/npcs/boy-1.png",
-    y: elderY,
-    x: elderX,
-    facing: Direction.DOWN,
-    canMove: false,
-    memory: {
-      metHero: false,
-    },
-    interactionHooks: [
-      {
-        id: "elder-rowan-greet",
-        type: "dialogue",
-        description: "Greet the elder",
-        payload: {
-          dialogueId: "elder-rowan-intro",
-        },
-      },
-    ],
-    actions: ["greet"],
-    metadata: {
-      archetype: "mentor",
-    },
-  });
-
-  const npcs: NPC[] = [elder];
-
-  // Carve an opening at the left wall ~3 tiles down from top to lead to the bluff passageway
-  if (tiles[3]?.[0] !== undefined) {
-    tiles[3][0] = FLOOR;
-    subtypes[3][0] = [TileSubtype.ROOM_TRANSITION];
-  }
-
-  return {
-    id: "story-outdoor-clearing",
-    mapData: { tiles, subtypes, environment: "outdoor" },
-    entryPoint,
-    transitionToPrevious,
-    entryFromNext: [exteriorDoorY, doorX],
-    transitionToNext: [doorY, doorX],
-    otherTransitions: [
-      {
-        roomId: "story-torch-town" as RoomId,
-        position: torchTownTransition,
-        returnEntryPoint: torchTownReturn,
-      },
-      // Opening on the left wall near the top leading to the bluff passageway
-      {
-        roomId: "story-bluff-passage" as RoomId,
-        position: [3, 0],
-        returnEntryPoint: [4, 1],
-      },
-    ],
-    npcs,
-  };
-}
-
-function buildOutdoorHouse(): StoryRoom {
-  const SIZE = 5;
-  const tiles: number[][] = Array.from({ length: SIZE + 2 }, () =>
-    Array.from({ length: SIZE + 2 }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: SIZE + 2 }, () =>
-    Array.from({ length: SIZE + 2 }, () => [] as number[])
-  );
-
-  for (let y = 1; y <= SIZE; y++) {
-    for (let x = 1; x <= SIZE; x++) {
-      tiles[y][x] = FLOOR;
-    }
-  }
-
-  const doorCol = Math.floor(SIZE / 2) + 1;
-  const entryPoint: [number, number] = [SIZE, doorCol];
-  const transitionToPrevious: [number, number] = [SIZE + 1, doorCol];
-  subtypes[transitionToPrevious[0]][transitionToPrevious[1]] = [
-    TileSubtype.DOOR,
-    TileSubtype.ROOM_TRANSITION,
-  ];
-  const entryFromNext: [number, number] = [SIZE - 1, doorCol];
-
-  const windowCols = [2, SIZE - 1];
-  for (const col of windowCols) {
-    const wallCol = col;
-    if (tiles[0]?.[wallCol] === WALL) {
-      subtypes[0][wallCol] = [TileSubtype.WINDOW];
-    }
-  }
-
-  const cornerTorches: Array<[number, number]> = [
-    [1, 1],
-    [1, SIZE],
-    [SIZE, 1],
-    [SIZE, SIZE],
-  ];
-  for (const [ty, tx] of cornerTorches) {
-    if (tiles[ty]?.[tx] === FLOOR) {
-      subtypes[ty][tx] = [TileSubtype.WALL_TORCH];
-    }
-  }
-
-  const caretaker = new NPC({
-    id: "npc-grounds-caretaker",
-    name: "Caretaker Lysa",
-    sprite: "/images/npcs/girl-1.png",
-    y: Math.max(2, Math.floor(SIZE / 2)),
-    x: Math.max(2, Math.floor(SIZE / 2)),
-    facing: Direction.DOWN,
-    canMove: false,
-    memory: {
-      sharedStories: 0,
-      restingSpotUnlocked: false,
-    },
-    interactionHooks: [
-      {
-        id: "lysa-intro",
-        type: "dialogue",
-        description: "Hear Lysa's introduction",
-        payload: {
-          dialogueId: "caretaker-lysa-intro",
-        },
-      },
-      // {
-      //   id: "lysa-chat",
-      //   type: "dialogue",
-      //   description: "Listen to Lysa's song",
-      //   payload: {
-      //     dialogueId: "caretaker-lysa-default",
-      //   },
-      // },
-
-      // {
-      //   id: "lysa-rest",
-      //   type: "custom",
-      //   description: "Request a moment of rest",
-      //   payload: {
-      //     action: "rest",
-      //   },
-      // },
-    ],
-    actions: ["talk", "rest"],
-    metadata: {
-      archetype: "caretaker",
-    },
-  });
-  const npcs = [caretaker];
-
-  return {
-    id: "story-outdoor-house",
-    mapData: { tiles, subtypes, environment: "house" },
-    entryPoint,
-    transitionToPrevious,
-    entryFromNext,
-    npcs,
-  };
-}
-
-function buildTorchTown(): StoryRoom {
-  const SIZE = 35;
-  // Initialize as walls so nothing outside the city walls renders as grass
-  const tiles: number[][] = Array.from({ length: SIZE }, () =>
-    Array.from({ length: SIZE }, () => WALL)
-  );
-  const subtypes: number[][][] = Array.from({ length: SIZE }, () =>
-    Array.from({ length: SIZE }, () => [] as number[])
-  );
-
-  const ensureSubtype = (y: number, x: number, subtype: TileSubtype) => {
-    const cell = subtypes[y][x] ?? [];
-    if (!cell.includes(subtype)) {
-      subtypes[y][x] = [...cell, subtype];
-    }
-  };
-
-  const grassMargin = 2;
-  const wallThickness = 2;
-  const wallMin = grassMargin;
-  const wallMax = SIZE - grassMargin - 1;
-
-  for (let layer = 0; layer < wallThickness; layer++) {
-    const top = wallMin + layer;
-    const bottom = wallMax - layer;
-    for (let x = wallMin; x <= wallMax; x++) {
-      tiles[top][x] = WALL;
-      subtypes[top][x] = [];
-      tiles[bottom][x] = WALL;
-      subtypes[bottom][x] = [];
-    }
-    for (let y = wallMin; y <= wallMax; y++) {
-      const left = wallMin + layer;
-      const right = wallMax - layer;
-      tiles[y][left] = WALL;
-      subtypes[y][left] = [];
-      tiles[y][right] = WALL;
-      subtypes[y][right] = [];
-    }
-  }
-
-  // Carve interior floor inside the double walls
-  for (let y = wallMin + wallThickness; y <= wallMax - wallThickness; y++) {
-    for (let x = wallMin + wallThickness; x <= wallMax - wallThickness; x++) {
-      tiles[y][x] = FLOOR;
-      if (!subtypes[y][x]) subtypes[y][x] = [];
-    }
-  }
-
-  // Bottom-left entrance opening with a short corridor
-  const entryColumn = wallMin + 1;
-  const transitionRow = SIZE - 1;
-  const corridorRows = [
-    wallMax - 2, // need one tile of space for the entrance into the town
-    wallMax - 1,
-    wallMax,
-    transitionRow - 1,
-    transitionRow,
-  ];
-  for (const row of corridorRows) {
-    if (tiles[row]?.[entryColumn] !== undefined) {
-      tiles[row][entryColumn] = FLOOR;
-      subtypes[row][entryColumn] = [];
-    }
-  }
-  const spawnRow = transitionRow - 1; // just inside the map at the bottom of the corridor
-  ensureSubtype(transitionRow, entryColumn, TileSubtype.ROOM_TRANSITION);
-
-  const entryPoint: [number, number] = [spawnRow, entryColumn];
-  const transitionToPrevious: [number, number] = [transitionRow, entryColumn];
-  const entryFromNext: [number, number] = [entryPoint[0], entryPoint[1]];
-
-  const buildStructure = (
-    top: number,
-    left: number,
-    width: number,
-    height: number
-  ): [number, number] => {
-    for (let y = top; y < top + height; y++) {
-      for (let x = left; x < left + width; x++) {
-        tiles[y][x] = WALL;
-        subtypes[y][x] = [];
-      }
-    }
-    const doorRow = top + height - 1;
-    const doorCol = left + Math.floor(width / 2);
-    // Mark building door as both a door and a room transition entry point
-    subtypes[doorRow][doorCol] = [
-      TileSubtype.DOOR,
-      TileSubtype.ROOM_TRANSITION,
-    ];
-
-    return [doorRow, doorCol];
-  };
-
-  const innerMin = wallMin + wallThickness;
-  const innerMax = wallMax - wallThickness;
-
-  // Central plaza checkpoint roughly in the middle of the inner area
-  const centerY = Math.floor((innerMin + innerMax) / 2);
-  const centerX = Math.floor((innerMin + innerMax) / 2);
-  if (tiles[centerY]?.[centerX] === FLOOR) {
-    subtypes[centerY][centerX] = [TileSubtype.CHECKPOINT];
-  }
-
-  // Library: taller building, placed near upper third
-  const libraryWidth = 3;
-  const libraryHeight = 5;
-  const libraryTop =
-    innerMin + Math.max(2, Math.floor((innerMax - innerMin) / 6));
-  const libraryLeft =
-    innerMin + Math.floor((innerMax - innerMin - libraryWidth) / 2);
-  const libraryDoor = buildStructure(
-    libraryTop,
-    libraryLeft,
-    libraryWidth,
-    libraryHeight
-  );
-
-  // Store: shorter building, placed below library with more spacing
-  const storeWidth = 3;
-  const storeHeight = 3;
-  const storeTop = Math.min(
-    innerMax - storeHeight - 2,
-    libraryTop + libraryHeight + 8
-  );
-  const storeLeft = libraryLeft;
-  const storeDoor = buildStructure(
-    storeTop,
-    storeLeft,
-    storeWidth,
-    storeHeight
-  );
-
-  const homeLastNames = [
-    "Ashwood",
-    "Brightfield",
-    "Cindercrest",
-    "Dawnhollow",
-    "Emberline",
-    "Frostvale",
-    "Glimmerstone",
-    "Highridge",
-    "Ironwood",
-    "Juniperfell",
-  ];
-  const homeAssignments: Record<string, string> = {};
-  const homeWidth = 3;
-  const homeHeight = 2;
-  // Helper: check if a rectangular area is clear floor (no overlaps)
-  const isAreaFree = (top: number, left: number, w: number, h: number) => {
-    for (let y = top; y < top + h; y++) {
-      for (let x = left; x < left + w; x++) {
-        if (y < innerMin || y > innerMax || x < innerMin || x > innerMax)
-          return false;
-        if (tiles[y]?.[x] !== FLOOR) return false; // something already occupies this spot
-      }
-    }
-    return true;
-  };
-  // Base rows (structured), then apply small jitter so they aren't in straight lines
-  const baseRows: number[] = [];
-  for (let y = innerMin + 4; y <= innerMax - homeHeight - 4; y += 5)
-    baseRows.push(y);
-  const leftColBase = innerMin + 3;
-  const rightColBase = innerMax - homeWidth - 3;
-
-  const jitter = (range: number) =>
-    Math.floor(Math.random() * (2 * range + 1)) - range;
-  const candidates: Array<{ top: number; left: number }> = [];
-  for (const baseTop of baseRows) {
-    // Left column home
-    candidates.push({ top: baseTop, left: leftColBase });
-    // Right column home
-    candidates.push({ top: baseTop, left: rightColBase });
-  }
-
-  const homesToPlace = Math.min(10, homeLastNames.length);
-  let homesPlaced = 0;
-  for (const base of candidates) {
-    if (homesPlaced >= homesToPlace) break;
-    // Apply small jitter (±2 tiles) and clamp inside bounds
-    const dy = jitter(2);
-    const dx = jitter(2);
-    let top = Math.max(
-      innerMin + 2,
-      Math.min(innerMax - homeHeight - 2, base.top + dy)
-    );
-    let left = Math.max(
-      innerMin + 2,
-      Math.min(innerMax - homeWidth - 2, base.left + dx)
-    );
-    // Keep a little breathing room from the center plaza
-    const tooCloseToCenter =
-      Math.abs(top - centerY) <= 2 && Math.abs(left - centerX) <= 2;
-    if (tooCloseToCenter) continue;
-    if (!isAreaFree(top, left, homeWidth, homeHeight)) {
-      // Try a couple nearby fallbacks without randomness explosion
-      const fallbackSpots: Array<[number, number]> = [
-        [top, left + 1],
-        [top, left - 1],
-        [top + 1, left],
-        [top - 1, left],
-      ];
-      let placed = false;
-      for (const [fy, fx] of fallbackSpots) {
-        if (
-          fy >= innerMin + 2 &&
-          fy <= innerMax - homeHeight - 2 &&
-          fx >= innerMin + 2 &&
-          fx <= innerMax - homeWidth - 2 &&
-          !(Math.abs(fy - centerY) <= 2 && Math.abs(fx - centerX) <= 2) &&
-          isAreaFree(fy, fx, homeWidth, homeHeight)
-        ) {
-          top = fy;
-          left = fx;
-          placed = true;
-          break;
-        }
-      }
-      if (!placed) continue;
-    }
-    const door = buildStructure(top, left, homeWidth, homeHeight);
-    homeAssignments[`${door[0]},${door[1]}`] =
-      homeLastNames[homesPlaced] ?? `Family ${homesPlaced + 1}`;
-    homesPlaced++;
-  }
-
-  return {
-    id: "story-torch-town",
-    mapData: { tiles, subtypes, environment: "outdoor" },
-    entryPoint,
-    entryFromNext,
-    transitionToPrevious,
-    metadata: {
-      homes: homeAssignments,
-      buildings: {
-        libraryDoor,
-        librarySize: [libraryWidth, libraryHeight],
-        storeDoor,
-        storeSize: [storeWidth, storeHeight],
-        homeSize: [homeWidth, homeHeight],
-      },
-    },
-  };
-}
-
 export function buildStoryModeState(): GameState {
   const entrance = buildEntranceHall();
   const ascent = buildAscentCorridor();
   const sanctum = buildSanctum();
-  const outdoor = buildOutdoorWorld();
+  const outdoor = buildOutdoorClearing();
   const outdoorHouse = buildOutdoorHouse();
   const torchTown = buildTorchTown();
   const bluffPassage = buildBluffPassageway();
